最新的代码生成

+++ src/main/java/ssn/lmj/demo/db/dobj/CityDO.java
@@ -6,12 +6,12 @@ package ssn.lmj.demo.db.dobj;
  * Owner: Minjun Ling
  * Creator: Robot
  * Version: 1.0.0
- * Since: Sun Jun 24 12:40:39 CST 2018
+ * Since: Mon Jun 25 09:42:12 CST 2018
  * Table: city
  */
 public final class CityDO implements Serializable {
     private static final long serialVersionUID = 1L;
-    public Integer id; // 城市编号
+    public Long    id; // 城市编号
     public Integer provinceId; // 省份编号
     public String  cityName; // 城市名称
     public String  description; // 描述
