+++ src/main/resources/sqls/city.sql
@@ -2,7 +2,7 @@ CREATE DATABASE springbootdb;
 
 DROP TABLE IF EXISTS  `city`;
 CREATE TABLE `city` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT '城市编号',
+  `id` bigint(10) unsigned NOT NULL AUTO_INCREMENT COMMENT '城市编号',
   `province_id` int(10) unsigned  NOT NULL COMMENT '省份编号',
   `city_name` varchar(25) DEFAULT NULL COMMENT '城市名称',
   `description` varchar(25) DEFAULT NULL COMMENT '描述',
